requirements/dev.txt
-r base.txt
ipython
django-extensions
isort
Sphinx
django-debug-toolbar
pygraphviz